Change for api requirement

+++ lib/data/models/product_details_model.dart
@@ -40,17 +40,17 @@ class ProductDetailsData {
 
   ProductDetailsData(
       {this.id,
-        this.img1,
-        this.img2,
-        this.img3,
-        this.img4,
-        this.des,
-        this.color,
-        this.size,
-        this.productId,
-        this.createdAt,
-        this.updatedAt,
-        this.product});
+      this.img1,
+      this.img2,
+      this.img3,
+      this.img4,
+      this.des,
+      this.color,
+      this.size,
+      this.productId,
+      this.createdAt,
+      this.updatedAt,
+      this.product});
 
   ProductDetailsData.fromJson(Map<String, dynamic> json) {
     id = json['id'];
@@ -65,7 +65,7 @@ class ProductDetailsData {
     createdAt = json['created_at'];
     updatedAt = json['updated_at'];
     product =
-    json['product'] != null ? Product.fromJson(json['product']) : null;
+        json['product'] != null ? Product.fromJson(json['product']) : null;
   }
 
   Map<String, dynamic> toJson() {
@@ -97,7 +97,7 @@ class Product {
   String? discountPrice;
   String? image;
   int? stock;
-  double? star;
+  int? star;
   String? remark;
   int? categoryId;
   int? brandId;
@@ -108,21 +108,21 @@ class Product {
 
   Product(
       {this.id,
-        this.title,
-        this.shortDes,
-        this.price,
-        this.discount,
-        this.discountPrice,
-        this.image,
-        this.stock,
-        this.star,
-        this.remark,
-        this.categoryId,
-        this.brandId,
-        this.createdAt,
-        this.updatedAt,
-        this.brand,
-        this.category});
+      this.title,
+      this.shortDes,
+      this.price,
+      this.discount,
+      this.discountPrice,
+      this.image,
+      this.stock,
+      this.star,
+      this.remark,
+      this.categoryId,
+      this.brandId,
+      this.createdAt,
+      this.updatedAt,
+      this.brand,
+      this.category});
 
   Product.fromJson(Map<String, dynamic> json) {
     id = json['id'];
@@ -140,9 +140,8 @@ class Product {
     createdAt = json['created_at'];
     updatedAt = json['updated_at'];
     brand = json['brand'] != null ? Brand.fromJson(json['brand']) : null;
-    category = json['category'] != null
-        ? Category.fromJson(json['category'])
-        : null;
+    category =
+        json['category'] != null ? Category.fromJson(json['category']) : null;
   }
 
   Map<String, dynamic> toJson() {
@@ -209,10 +208,10 @@ class Category {
 
   Category(
       {this.id,
-        this.categoryName,
-        this.categoryImg,
-        this.createdAt,
-        this.updatedAt});
+      this.categoryName,
+      this.categoryImg,
+      this.createdAt,
+      this.updatedAt});
 
   Category.fromJson(Map<String, dynamic> json) {
     id = json['id'];
