+++ lib/data/models/product_model.dart
@@ -33,7 +33,7 @@ class ProductData {
   String? discountPrice;
   String? image;
   int? stock;
-  double? star;
+  int? star;
   String? remark;
   int? categoryId;
   int? brandId;
@@ -44,21 +44,21 @@ class ProductData {
 
   ProductData(
       {this.id,
-        this.title,
-        this.shortDes,
-        this.price,
-        this.discount,
-        this.discountPrice,
-        this.image,
-        this.stock,
-        this.star,
-        this.remark,
-        this.categoryId,
-        this.brandId,
-        this.createdAt,
-        this.updatedAt,
-        this.brand,
-        this.category});
+      this.title,
+      this.shortDes,
+      this.price,
+      this.discount,
+      this.discountPrice,
+      this.image,
+      this.stock,
+      this.star,
+      this.remark,
+      this.categoryId,
+      this.brandId,
+      this.createdAt,
+      this.updatedAt,
+      this.brand,
+      this.category});
 
   ProductData.fromJson(Map<String, dynamic> json) {
     id = json['id'];
@@ -77,7 +77,7 @@ class ProductData {
     updatedAt = json['updated_at'];
     brand = json['brand'] != null ? Brand.fromJson(json['brand']) : null;
     category =
-    json['category'] != null ? Category.fromJson(json['category']) : null;
+        json['category'] != null ? Category.fromJson(json['category']) : null;
   }
 
   Map<String, dynamic> toJson() {
@@ -144,10 +144,10 @@ class Category {
 
   Category(
       {this.id,
-        this.categoryName,
-        this.categoryImg,
-        this.createdAt,
-        this.updatedAt});
+      this.categoryName,
+      this.categoryImg,
+      this.createdAt,
+      this.updatedAt});
 
   Category.fromJson(Map<String, dynamic> json) {
     id = json['id'];
